+++ sme/src/sme_module.cpp
@@ -38,8 +38,27 @@ void pybindModule(pybind11::module &m) {
             Model: the spatial model
         )");
   m.def("open_example_model", openExampleModel,
+        pybind11::arg("name") = "very-simple-model",
         R"(
         opens a built in example spatial model
+
+        The model name can optionally be specified to one of the following built
+        in models:
+
+        * "ABtoC"
+        * "brusselator-model"
+        * "circadian-clock"
+        * "gray-scott"
+        * "liver-simplified"
+        * "liver-cells"
+        * "single-compartment-diffusion"
+        * "very-simple-model"
+
+        If the name is not specified the the default
+        example model is "very-simple-model".
+
+        Args:
+            name (str, optional): name of the example model to open, default: "very-simple-model"
 
         Returns:
             Model: the example spatial model
@@ -67,14 +86,16 @@ Model openFile(const std::string &filename) { return Model(filename); }
 
 Model openSbmlFile(const std::string &filename) { return Model(filename); }
 
-Model openExampleModel() {
+Model openExampleModel(const std::string &name) {
   Model m;
   std::string xml;
-  if (QFile f(":/models/very-simple-model.xml");
+  if (QFile f(QString(":/models/%1.xml").arg(name.c_str()));
       f.open(QIODevice::ReadOnly | QIODevice::Text)) {
     xml = f.readAll().toStdString();
   } else {
-    throw SmeInvalidArgument("Failed to open example model");
+    throw SmeInvalidArgument("Failed to open example model '" + name +
+                             "'. Type help(sme.open_example_model) to see the "
+                             "available built-in models.");
   }
   m.importSbmlString(xml);
   return m;