+++ sme/src/sme_module.hpp
@@ -9,6 +9,6 @@ void pybindModule(pybind11::module &m);
 
 Model openFile(const std::string &filename);
 Model openSbmlFile(const std::string &filename);
-Model openExampleModel();
+Model openExampleModel(const std::string &name);
 
 } // namespace sme